process of creating a project
Relations between MainMenuFrame, ButtonHolder and CreateProjectFrame has been done by adding ActionListeners.

When clicked on the "CreateProject" button on the ButtonHolder there will be a CreateProjectFrame initialized. When it is done a Project instance will be created and passed to the User. Then there will be a button added to the ButtonHolder.